requirements.txt
fastmcp>=2.0.0
python-jenkins>=1.8.0
pydantic>=2.0.0
pyyaml>=6.0
